refactor: apply result both in failure and success scenarios

+++ TaskManagerAPI.Application/UseCases/DeleteTask.cs
@@ -1,20 +1,27 @@
-﻿using TaskManagerAPI.Domain.Interfaces;
+﻿using System.Reflection.Metadata.Ecma335;
+using TaskManagerAPI.Domain.Common.Results;
+using TaskManagerAPI.Domain.Interfaces;
 
 namespace TaskManagerAPI.Application.UseCases;
 
 public interface IDeleteTask
 {
-    Task ExecuteAsync(Guid id);
+    Task<Result> ExecuteAsync(Guid id);
 }
 
 public class DeleteTask(ITaskRepository repository) : IDeleteTask
 {
-    readonly ITaskRepository _repository = repository;
+    private readonly ITaskRepository _repository = repository;
 
-    public async Task ExecuteAsync(Guid id)
+    public async Task<Result> ExecuteAsync(Guid id)
     {
-        var item = await _repository.GetByIdAsync(id) ?? throw new Exception("Not Found");
+        var item = await _repository.GetByIdAsync(id); 
+        
+        if (item is null)
+            return Result.Failure(TaskManagementErrors.NotFound(id));
 
         await _repository.DeleteAsync(item);
+
+        return Result.Success();
     }
 }
